+++ monix/src/main/scala/outwatch/ext/monix/Handler.scala
@@ -1,0 +1,9 @@
+package outwatch.reactive.handlers
+
+import _root_.monix.reactive.{Observable, Observer}
+
+import colibri._
+import colibri.ext.monix._
+import outwatch.reactive._
+
+object monix extends HandlerEnvironment[Observer, Observable, MonixSubject, MonixProSubject]
