+++ reactive/src/main/scala/outwatch/reactive/Handler.scala
@@ -1,0 +1,5 @@
+package outwatch.reactive
+
+import colibri._
+
+object handler extends HandlerEnvironment[Observer, Observable, Subject, ProSubject]
