+++ reactive/src/main/scala/outwatch/reactive/HandlerEnvironment.scala
@@ -2,7 +2,7 @@ package outwatch.reactive
 
 import colibri._
 
-trait HandlerEnvironment[SinkT[-_], SourceT[+_], HandlerT[_], ProHandlerT[-_, +_]] {
+class HandlerEnvironment[SinkT[-_] : LiftSink, SourceT[+__] : LiftSource, HandlerT[_] : CreateSubject, ProHandlerT[-_,+_] : CreateProSubject] {
 
   type HandlerSink[-T] = SinkT[T]
   type HandlerSource[+T] = SourceT[T]
@@ -10,20 +10,9 @@ trait HandlerEnvironment[SinkT[-_], SourceT[+_], HandlerT[_], ProHandlerT[-_, +_
   type Handler[T] = HandlerT[T]
   type ProHandler[-I, +O] = ProHandlerT[I,O]
 
-  val Handler: HandlerFactory[Handler]
-  val ProHandler: ProHandlerFactory[ProHandler]
+  val Handler: HandlerFactory[Handler] = new HandlerFactory[Handler]
+  val ProHandler: ProHandlerFactory[ProHandler] = new ProHandlerFactory[ProHandler]
 
-  val HandlerSink: SinkFactory[HandlerSink]
-  val HandlerSource: SourceFactory[HandlerSource]
-}
-
-object HandlerEnvironment {
-  def apply[SinkT[-_] : LiftSink, SourceT[+__] : LiftSource, HandlerT[_] : CreateSubject, ProHandlerT[-_,+_] : CreateProSubject]: HandlerEnvironment[SinkT, SourceT, HandlerT, ProHandlerT] =
-    new HandlerEnvironment[SinkT, SourceT, HandlerT, ProHandlerT] {
-      val Handler = new HandlerFactory[Handler]
-      val ProHandler = new ProHandlerFactory[ProHandler]
-
-      val HandlerSink = new SinkFactory[HandlerSink]
-      val HandlerSource = new SourceFactory[HandlerSource]
-    }
+  val HandlerSink: SinkFactory[HandlerSink] = new SinkFactory[HandlerSink]
+  val HandlerSource: SourceFactory[HandlerSource] = new SourceFactory[HandlerSource]
 }
