--- reactive/src/main/scala/outwatch/reactive/package.scala
@@ -1,7 +1,0 @@
-package outwatch
-
-import colibri._
-
-package object reactive {
-  val handler = HandlerEnvironment[Observer, Observable, Subject, ProSubject]
-}
