+++ tests/src/test/scala/outwatch/DomEventSpec.scala
@@ -9,7 +9,7 @@ import outwatch.Deprecated.IgnoreWarnings.initEvent
 
 import outwatch.dsl._
 import colibri.ext.monix._
-import outwatch.ext.monix.handler._
+import outwatch.reactive.handlers.monix._
 
 import outwatch.util.LocalStorage
