+++ tests/src/test/scala/outwatch/OutWatchDomSpec.scala
@@ -16,7 +16,7 @@ import outwatch.dsl._
 import outwatch.helpers._
 import outwatch.interpreter._
 import colibri.ext.monix._
-import outwatch.ext.monix.handler._
+import outwatch.reactive.handlers.monix._
 import outwatch.reactive.handler.{Handler => Internal}
 import colibri._
