+++ tests/src/test/scala/outwatch/SnabbdomSpec.scala
@@ -7,7 +7,7 @@ import outwatch.Deprecated.IgnoreWarnings.initEvent
 
 import outwatch.dsl._
 import colibri.ext.monix._
-import outwatch.ext.monix.handler._
+import outwatch.reactive.handlers.monix._
 
 import snabbdom._
